refactor: components separation + dialogs lazy loading
