+++ src/components/viewer/delete-dialog.tsx
@@ -1,0 +1,56 @@
+"use client";
+
+import { useTranslations } from "next-intl";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "~/components/ui/alert-dialog";
+
+export interface DeleteDialogProps {
+  open: boolean;
+  onClose: () => void;
+  fileName: string | undefined;
+  onConfirm: () => void;
+}
+
+export default function DeleteDialog({
+  open,
+  onClose,
+  fileName,
+  onConfirm,
+}: DeleteDialogProps) {
+  const t = useTranslations("Viewer");
+
+  return (
+    <AlertDialog open={open} onOpenChange={(open) => !open && onClose()}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>{t("delete.title")}</AlertDialogTitle>
+          <AlertDialogDescription>
+            {fileName && (
+              <>
+                {t("delete.description")}
+                <br />
+                <br />
+                <span className="font-medium text-foreground">{fileName}</span>
+              </>
+            )}
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>{t("delete.cancel")}</AlertDialogCancel>
+          <AlertDialogAction variant="destructive" onClick={onConfirm}>
+            {t("delete.confirm")}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
